Refactor imports for @hedvig-ui/TextArea
